Create pipe.c
A parent process creates a pipe for communication with its child. The parent process takes a txt file name as input from the user and writes the contents of a file, one line at a time, into the pipe. The child process reads the pipe and prints the output on the screen.