Add files via upload
The parent reads text input from the user, converts all text into uppercase, and writes it to the screen. The child process reads input from standard input and writes the output with line number information to standard output. Use dup2() to pipe parent output to the child such that the child process outputs line-numbered uppercase text on the screen.