Add files via upload
Simulating a process execution by generating a large string of address references. These are written out to a file.

You will write a second process that simulates memory management. It reads the file and, if the needed page is in memory, calculates the physical address and writes the logical and physical addresses back into the file. If the page is not present in memory, it will load and then mark the address as a page fault.

You will write a third program which reads the updated file containing information on physical and logical addresses, and page faults and generates certain statistics.