Add files via upload
Will be extending the process simulator and memory manager written in the last lab. We will now handle the case where the total number of pages required for the processes exceeds the number of frames in physical memory. We Will need to write a function that takes the PID of the process making a page request, and other data structures required for the different page replacement algorithms and return the frame number in which the new page shall be loaded.